# backend/internal/transport/http/static/openapi.yaml
openapi: 3.0.3
info:
  title: Cash Factories News API
  version: 0.1.0
  description: >
    API for collecting real news and exposing rewritten versions in different moods
    without changing facts.
servers:
  - url: /
    description: Current host
paths:
  /health:
    get:
      summary: Health check
      tags:
        - System
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
  /api/v1/news:
    get:
      summary: List news
      tags:
        - News
      parameters:
        - in: query
          name: mood
          required: false
          schema:
            $ref: "#/components/schemas/Mood"
          description: Filter by rewrite mood
      responses:
        "200":
          description: News list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/News"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/news/rewrite:
    post:
      summary: Rewrite a selected news item in the requested mood
      tags:
        - News
      parameters:
        - in: query
          name: external_id
          required: true
          schema:
            type: string
          description: External article identifier from the source
        - in: query
          name: mood
          required: false
          schema:
            $ref: "#/components/schemas/Mood"
          description: Mood used for rewriting
      responses:
        "200":
          description: Rewritten news item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: News not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/news/{id}:
    get:
      summary: Get single news item
      tags:
        - News
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: News item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/News"
        "404":
          description: News not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/news/sync:
    post:
      summary: Sync latest news and rewrite them in selected mood
      tags:
        - News
      parameters:
        - in: query
          name: mood
          required: false
          schema:
            $ref: "#/components/schemas/Mood"
          description: Mood used for rewriting
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 10
          description: Number of latest news items to fetch and rewrite
      responses:
        "202":
          description: Sync job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJob"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/v1/jobs/{id}:
    get:
      summary: Get sync job status
      tags:
        - Jobs
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync job state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJob"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    Mood:
      type: string
      enum:
        - neutral
        - happy
        - sad
        - ironic
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
      required:
        - status
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: sync completed
        count:
          type: integer
          example: 10
      required:
        - status
        - count
    SyncJob:
      type: object
      properties:
        ID:
          type: string
        Status:
          type: string
          enum:
            - pending
            - running
            - completed
            - failed
        Mood:
          $ref: "#/components/schemas/Mood"
        Limit:
          type: integer
        ProcessedCount:
          type: integer
        TotalCount:
          type: integer
        Error:
          type: string
        CreatedAt:
          type: string
          format: date-time
        UpdatedAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    News:
      type: object
      properties:
        ID:
          type: string
        Title:
          type: string
        OriginalText:
          type: string
        RewrittenText:
          type: string
        Mood:
          $ref: "#/components/schemas/Mood"
        SourceName:
          type: string
        SourceURL:
          type: string
          format: uri
        PublishedAt:
          type: string
          format: date-time
        CreatedAt:
          type: string
          format: date-time
        ExternalID:
          type: string
        FactChecksum:
          type: string
        OriginalDigest:
          type: string